+++ bundlers/undeclared-identifiers.d.ts
@@ -1,5 +1,5 @@
 ﻿// https://github.com/goto-bus-stop/undeclared-identifiers
 declare module "undeclared-identifiers" {
-    function undeclaredIdentifiers(source: any, opts: { properties?: any; wildcard?: any;[key: string]: any }): { identifiers: string[]; properties: any[] };
+    function undeclaredIdentifiers(source: any, opts: { properties?: any; wildcard?: any; [key: string]: any }): { identifiers: string[]; properties: any[] };
     export = undeclaredIdentifiers;
 }